use a full GOPATH in testdata
We can't use the existing GOPATH with testdata as that makes for a
non-importable path, which confuses go/build. With this change, we can
fully use go/build to replace some of our importing logic.

While at it, sort the ProtoFile slice from the request to ensure
deterministic output yet again. Occasionally, the "Empty from public
import" part would be omitted if the map iteration was weird.