loader: handle int, bool, array and struct
The Gunk example now works.

'protoType' can now return a 'label', this is useful for any fields that
need to be 'repeated', 'optional' or 'required' in proto.

'protoType' handles bool, int, int32, array and struct. Array's are now
possible because a 'label' is returned. For struct's and enum's, we check whether
the 'enum' or 'message' have been parsed and are in a valid gunk file.
This is likely not the most robust way, but works for now.

Added a slightly more useful error message when 'protoc-gen-go' fails
that includes and stdout and stderr output.

Added the command to run the example in the readme.