loader: don't panic on ./example/util
In master, trying to generate the example seems to panic:

	$ go install
	$ gunk ./examples/util
	panic: unknown or mismatched type: type: loader.httpRule, wire type: varint

The underlying reason is that we rolled our own type and definition for
Google's http protobuf extension. That worked before, but has broken at
some point in the past few months, likely during a protobuf or protobuf
generator release.

To fix the issue and ensure this doesn't happen again, simply use the
protobuf-generated code from google.golang.org to set up their
extensions. This is simpler, and should work as long as we update all
protobuf deps in sync.

Running gunk on the example works on my machine now.