re-enable the tests
Now that we are a module and support modules, this is trivial to do.

First, we no longer need GOPATH, so testdata/src/util/foo.gunk can be
just testdata/foo.gunk, with a go.mod for the module testdata.tld/util.

Second, we don't need to copy github.com/gunk/opt anymore, as the go
tools will download it as necessary. Checking what it generates isn't
useful either, as we don't really care.

Third, we can test both the generation and the build of the generated
files even in -short, thanks to GOCACHE. A cold 'go test' may take up to
six seconds, but subsequent invocations take half a second on my
machine.