loader: make full use of go/types
For some reason, we were being silly and inspecting the syntax tree to
translate types from Gunk to protobuf. This is much simpler and robust
with go/types.

Also solves the TODO, since we no longer have to resolve SelectorExpr
nodes ourselves.

Otherwise, nothing is changed. Added more field types to testdata to
ensure that we didn't break a thing.

Fixes #12.