loader: add map support
Added support for the Go map type to the Protobuf map representation.

The Protobuf map representation is a new 'NestedMessage' on the parent
message containing the map. The nested message is a message that contains
two fields; 'key' and 'value ('map[key]value'). An option needs to be
set on the nested message to indicate that this is a message that
represents a map.