testdata: start adding error test cases
For example, what happens if we have a Go package with zero gunk files?

Use github.com/rogpeppe/go-internal/testscript for this, which pulls out
cmd/go's script testing features. This is useful to have small and
maintainable test cases without needing dozens of directories and files
scattered on the filesystem.

This is the initial setup with the first simple test script; more will
follow.

While at it, move testdata/* to testdata/generate/*, to separate the
files required by different tests.

Fixes #36.