loader: preparation for custom proto package names
Initially I wanted to get this done independently in a single commit,
but it turns out to be trickier. We need to know all protobuf package
names upfront, so that we can write fully qualified names properly.

I had another PR in the works that revamped package loading and
introduced the notion of a "gunk package", so that should be merged to
help with this task. So it's best to merge this piece first, and then
finish off the feature once the loading refactor is merged.

Updates #44.