config: added basic gunkconfig
Added a config package to find and read a .gunkconfig. It behaves in a
similar way to go mod, which starts in the working directory and then
moves up each parent checking for a .gunkconfig.

The .gunkconfig can contain one or more 'generate' sections. Each
generate section is a single command to be run. Currently, the command
is assumed to be a protoc-gen-* command, which takes in the proto
filedescriptor as stdin and outputs a proto filedescriptor as stdout.