config,generate: use protoc to generate other language proto bindings
Use protoc to generate the other language bindings for now. We can use
protoc to generate most other languages, and for those protoc can't
generate, we will need to use protoc-gen-* plugins.

This PR adds a new config paramater to gunkconfg, 'protoc=<language>'.

'protoc' is more strict than 'protoc-gen-go'; If an enum is found without
any values, we need ignore the enum. If a typeName to a message field is
empty, we need to set the typeName to 'nil', rather than a pointer to
an empty string.

Fixes #46