generate, mod: copy pointer values before updating them
In each 'generateProtoc' run we modify the protofile name, however, this
was a pointer, and was changing the value for all runs of
'generateProtoc' and 'generatePlugin'. Make a dumb copy of the protofile
before we use it.

Updated the ini parser to a forked version that handles multiple
section names, and getting keys from the different sections, correctly.