generate: change protoc generator output location
To keep functionality the same as protoc-gen-* generators, the protoc
generators have been changed to output the generated proto files in the
same directory as the .gunk file.

In order to tell protoc where to output the files, we need to change the
proto file name to be the location on disk we want it to write to. Since
dependencies are based on proto file names, all dependencies also need
to be changed to be the location on disk.

Fix some issues with maps and empty string TypeName.