config,generate: out path is configurable in gunkconfig
If an 'out' key is specified in a gunkconfig generate section, that will
be used as the directory to output the generated files. If the key
doesn't exist, we default to outputting the generated file to where the
'*.gunk' file was found (the current behaviour).

Stop running 'protoc-gen-go' by default. This didn't work with having
configurable out paths, because each out path is seperate for each
section, and there is no good way to set the out path for the Go
generator. Instead update the .gunkconfig to run the Go generator.