format: add missing proto sequence numbers to a struct
If a struct field is missing the proto sequence number, attempt to add
the sequence number to the tag if it exists, or create a new tag and add
it to that.

However, there is no easy way to update the struct field tag if the 'pb'
tag is already present (whether empty, or containing a non-number). For now,
we return an error to either remove the tag, or add the missing sequence number.

Closes #74