config: remove default config and assume .git is project root
Removed the default config. If 'gunk generate' is run and no
'.gunkconfig' is found, an error is now showed to the user. 'gunk
generate' can now only be run if a .gunkconfig file is found.

Previously, we assumed the project root was where a  'go.mod' file was
found. This has been updated to now also look for a '.git' file or
directory.

Closes #101