loader: perform all gunk loading here
We were using the loader for two purposes. The simpler was the
formatter's use case, where we only need to parse the given packages.

The second and more complex is the generator's. It needs to parse the
given Gunk packages, all of their Gunk dependencies, and type-check all
of them including their gunk tags.

Before, all this extra work was done in the generator package. This
worked okay, but led to complex lazy loading and type-checking all over
the place.

Instead, have the loader package mirror go/packages more closely. By
default it will only do what the formatter needs. But if Types==true, it
will also typecheck the Gunk packages and their gunk tags, and
consequently do the same for all their transitive Gunk dependencies.

All the loading is entirely in the loader package, and the generate
package no longer evaluates tags in a handful of places.

Fixes #112.