all: bundle some proto libraries with gunk
Bundle the following proto libraries with Gunk:

- google/api/annotations.proto
- google/protobuf/empty.proto
- google/protobuf/timestamp.proto
- google/protobuf/duration.proto

these are the libraries that are currently used by Gunk, and require
calling out to protoc to load from disk. These will now be included in
Gunk.

The included libraries are being stored in their FileDescriptorProto
on-the-wire format. Go generate is used to create these on-the-wire
formats of the libraries, which runs 'protoc' to load the libraries from
disk and write the on-the-wire formats to the 'generated/' directory.

Currently using 'shurcooL/vfsgen' to create gzipped contents of the
files and store them in memory.

Fixes #116