testdata: use a static GOPROXY for the scripts
A 'go test' with the cache warm now takes ~0.7s instead of ~4s on my
laptop.

The command used to "vendor" the module in this static GOPROXY was
github.com/rogpeppe/go-internal/cmd/txtar-addmod:

	txtar-addmod -all testdata/mod github.com/gunk/opt@latest

Finally, we need to add the same version requires to each test go.mod
file, so that we don't have to figure out which version is latest at
each test run.

Fixes #119.