config: allow shortened generate sections for protoc-gen-*
'protoc-gen-*' generators can now be used as shortened generate sections
in the gunkconfig. Previously, we just assumed that any
'[generate <x>]' in the gunkconfig should be passed through as 'proto --<x>_out'.
Now, we can have '[generate go]' and this will check for a 'protoc-gen-go' binary,
if it exsists that is used, otherwise it falls back to being passed
through to protoc.

Fixes #96