all: run tests in parallel
The script tests were running in parallel among themselves already,
which is built into testscript. However, they weren't running in
parallel to TestGenerate.

TestGenerate and some of the script tests are the slowest without maxing
out on cpu, so this reduces 'go test' time on my laptop on battery from
~2.4s to ~1.4s.

TestGenerate writes to files in testdata/generate, but that's still fine
for t.Parallel - since that means it can run in parallel with other
parallel tests, but not itself.