dump: command that outputs filedescriptorset
Added a new 'dump' command that outputs a single Gunk package's
FileDescriptorSet (as specified in descriptor.proto). This should help
when trying to debug the differences between what Gunk and Protoc
generate.

The dump command will output in 'raw' format as default (the same as
protoc -o), but can be specified to dump as 'json' format. The output
will be sent to stdout by default, but can also be written to a file.

Fixes #136