generate: sort protofile dependencies
Attempt to sort the FileDescriptorSet in an order such that any proto
files without dependencies are first in the list, and the rest are at
the end.

Use a fairly dumb, but simple approach to determine which protofile's
depend on others. We can use this approach because we don't load that
many protofile's, and the ones we do load don't depend on each other.

This is required to fix an issue with the protoc nodejs grpc generator,
which seems to expect the protofiles to be in an order where any
depdendent protofile's are earlier on in the list.

Fixes #137