README: minor edits to be less verbose
Remove the "required installs" step. We can easily merge the Go
dependency with the install step. And the protoc dependency will be
dropped soon, as gunk will download it at run-time if necessary.

Pointing at the issue tracker is also not terribly useful, as GitHub is
very ubiquitous these days.

Finally, cut the help output, as we don't want to keep that up to date
forever, and we don't want the README to be that long. The user can
easily get the help text themselves if needed.

And some other minor adjustments, like removing an unnecessary .git
extension and joining the GO111MODULE line with another.