convert: proto fileoptions
Updated convert to convert proto 'FileOptions' to the corresponding Gunk
annotations.

While doing this, it was required to keep a record of which
imports of Gunk annotations are being used. Now we only output imports
that are actually in use, or no imports at all if there are no imports.

Updates #143