generate: don't include stdout when an exec errors
Most programs, upon erroring, will only include the reason in stderr.
This led to some confusing error messages on common exec errors, such as
the binary missing in $PATH:

	error: error executing protoc-gen-print: , exec: "protoc-gen-print": executable file not found in $PATH

To fix this, don't include stdout in error messages. But do include
stderr if it wasn't already written out directly because of the verbose
flag.

Also make the log package output to os.Stderr by default, making it
configurable. Logging should generally go to stderr, as stdout is
reserved for the result or output of a program or command. This also
means that a plugin's stderr will always go to stderr if it errors, even
if gunk is run in verbose mode.

Finally, clean up the scripts files to not mix tabs and spaces
unnecessarily.