generate: refactor how we sort proto files
First, add a test script that uses a custom Go generator which checks
that the ProtoFile slice is sorted topologically. We've found in the
past that some generators require that, so the easiest way to add a
regression test is to implement a dummy generator which checks that
property.

Second, redo the topological sort to support all inputs. The old code
only split the list of files in two chunks, which may quickly break if
one has multiple levels of package dependencies.

There are efficient algorithms we could have chosen which would have
taken over fifty lines of code, but a much simpler quadratic algorithm
will do the job for the foreseeable future in under twenty.

Finally, remove the sorts of files by name. We no longer have multiple
proto files per Gunk package, so we don't care about the order of
FileToGenerate. And ProtoFile is sorted topologically, so the name
sorting was useless.