convert: include filename in errors
Add the filename to error messages in gunk convert. This is now required
because gunk convert can handle more than a single proto file.

Change the pkgOpts to use proto.Option rather than a custom struct so
that we don't lose positional information about an option, and can error
correctly on that option. Also remove 'pkgOpt' as it is no longer used.

Fixes #154