generate: discard empty doc comments
If a gunk comment only contained a gunk tag, don't keep it around as a
documentation comment in the generated code.

Add a test, and update generate_options.txt to be more permissive with
the whitespace. Since we removed a few bad comments there, gofmt now
aligns the field types, so a few spaces have been introduced.

Fixes #110.