main: force kingpin not to use os.Exit directly
We implement a main1 function which returns exit codes, so that
testscript can use it as a builtin command. If the function exits
prematurely, testscript will fail to do post-cmd work like writing
profiles to disk.

Write the code to handle that properly, and make CI produce a coverage
profile to ensure it's not broken again. We don't use the profile just
yet, but we might in the future.

While at it, start testing on 1.12 on CI too, and update the module
dependencies.

Fixes #171.