log: simplify package
The log package only needs to care about printing lines, and whether
some lines are verbose or not. We also have a wrapper for exec.Cmd which
is handy. Everything else is redundant.

This can be shown by how halving the number of APIs in the log package
doesn't make the rest of the code any more complex. If anything, the
code using the log package is now simpler to follow.