generate: greatly simplify file descriptor sets
Generators directly within protoc insist on writing the files
themselves, so we must give it a CodeGeneratorRequest that has been
modified to contain the right paths.

However, we were doing this in a very roundabout way. We did it for all
generators, and we used a map, which is unnecessary.

The actual logic only requires about ten lines of code in
generateProtoc. We can remove over fifty lines, thanks to that.

The other piece that the old code did was replace ".gunk" with ".proto".
However, all packages are generated as a single "all.proto" file now, so
that line had been a no-op for a while.