loader: rethink package loading errors
go/packages only uses its error return for fatal errors, such as when
the query is malformed or when it cannot call 'go list'.

All other errors, such as typecheck errors, are attached to packages.
This is because loading can often continue in the face of these
non-fatal errors, and it's useful for tools to report many errors at
once.

Do the same in our loader package. On the upside, this removes
ValidateError, removes a few error returns, and consolidates the API
with go/packages.

The only disadvantage is that we now need to port PrintErrors from
go/packages, which also requires Visit. But it's likely we were going to
need Visit sooner or later, anyway.