loader: named imports and message options
When loading and converting a proto file to a gunk file, use named
imports for imports that are likely to have the same package name.

The first time an import is seen, that gets to keep its package name.
The next time an import with that package name is encounted, that
is given a named import.

Added convert options for a proto message type.

Updates: #143