loader: set up Name and ProtoName when loading files
We were leaving pkg.Name as a TODO when loading single files. And this
was usually fine, except when generating lone Gunk files under some
circumstances.

Cédric ran into one of those in the added test script, which he
supplied. Fix the bug in the loader by grabbing the package name from
all the files, much like cmd/go does.

Finally, remove the ProtoName=="" workaround that was added recently
because of this bug.