testdata/mod: fix the module copies
gunk/opt was invalid, as reported by sum.golang.org; go.mod had been
tampered somehow. Re-add it.

We also need to be able to resolve github.com/google/protobuf. The
module is massive, and txtar-addmod would add ~1.5MiB in a file.
Fortunately, we just need to resolve the module, and not build it.

So a stub txtar file is enough. Turn off sum.golang.org, as it
(correctly) reports that the source is different. We don't care.

With these changes, 'go test' passes on 'go version devel +d97bd5d07a
Sat May 25 22:38:01 2019 +0000 linux/amd64'.