reflectutil: move more reflect magic here
We can move a lot of work to this helper function. For example, grabbing
the name from an identifier key, the value from a literal, or unquoting
a value string.

Also make it support converting types, so we can assign an int32 value
to an enum field, since it has a named type.

We can also remove many uses of snaker; the field search was already
case-insensitive, so removing all underscores is all that we need to
make it work.

This means we can drastically simplify and unify many uses of SetValue,
making the meat of the code simpler and less verbose.