make protoc version, path configurable
Add a protoc config section to allow users to pin a specific protoc
version. Make the path configurable as well- gunk will attempt to
execute the binary at the path, or if none exists, gunk will download
the binary to this path.

Add a subcommand to the download command. The user now has to specify
that they want to download protoc. This makes the download command
more flexible, for when we want to use it to download other
protoc-related tools. Allow users to specify the path and version
with the download command as well.

Update the README to reflect the changes and fix miscellaneous
formatting problems while at it.

[OB-367] #done