generate: replace field names from UpperCamelCase to snake_case
We used UpperCamelCase when writing variable/field names in Go, but the
convention for writing a protobuf file is to use a snake_case, such
as the following:

```
message SomeMessage {
    string user_id = 1
}
```

Currently, the `generate` code just copies the field name, which results
in this proto file before outputting the generated artifact:

```
message SomeMessage {
    string UserID = 1
}
```

which causes issues with the json tag on the file descriptor

This commit fixes the above by replacing the field name to snake_case,
per protobuf specification.

Fixes #263.