scopegen: create standardized scope challenges
Using a set of authorized scopes, ensure the request is Authorized by
testing against the method's defined scopes.

In order to ensure the method scopes are not changed during execution,
the method scopes are unexported.

Usage:  Embed the <package>.ServiceScope type in the service struct,
then use the `Any` or `All` challenges in an interceptor, with the set
of user authorized scopes.