docgen: write hostName and basePath to examples
We are using `basePath` wrong, at least in the testing files and docgen.
With host `openbank.com`, basepath `v1` and method path `v1/accounts/{AccountID}`,
the whole URL is `openbank.com/v1/v1/accounts/{AccountsID}`.

We are not using it like that in the templates.

Similarly, we are not using schemes correctly. Scheme should not be part of the hostname.

This commit fixes the issues by printing the paths correctly, and replacing them in the test scripts.
(And also changing `v1` to `path` in the basePath to make it less ambiguous.)

Codegen prints `https` or `http` if they are present, otherwise prints `wss` or `ws`, even when the CURL example is not correct. We can fix that later.

Fixes https://github.com/gunk/gunk/issues/282