generate: use a separate lock file when downloading protoc
Before this patch 'sudo rm -rf .cache && go test ./...' would result in
lots of errors like:

	error: error executing "[...]/protoc-v3.9.1": [...]: text file busy

This was because we were using the binary itself as the file being
locked. Because the lock was grabbed with write access, and many
operating systems like Linux forbid executing a binary that's open for
writing, we were seeing data races when downloading.

Here's a description of the chain of events that could lead to a race
between two tests, assuming the empty cache:

	Test A: finds protoc is missing, locks the binary for writing
	Test B: finds protoc is missing, locks the binary for writing
	Test A: wins the race for the lock, begins downloading protoc
	Test B: loses the race for the lock, blocks
	Test A: finishes the download, closes the file, and releases the lock
	Test B: unblocks and grabs the lock, opening the file for writing
	Test A: tries to verify that protoc works, executing the binary
	Test A: errors, since Test B has the file open for writing

There are other cases where we could hit such an error, too. The above
is just one of the examples. To fix this problem, use a separate lock
file, as explained in the added comments.

The tests now reliably pass on Go 1.14, regardless of whether the cache
file exists or not.