docgen: do not generate "annex" (no services file)
Currently, we generate "annex file" for gunk files without services

This file is, however, basically useless; nobody wants to read all sub-types without context, and all
the messages should be already documented in the services.

And if we have any types, not "reachable" from services, they should not be generated at all anyway.

I have changed some of the tests that depended on Annex and slightly simplify them.

Note that with this change, we now don't need Messages in the File parse model at all!

Also note that this uncovered some bug in the JSON example generation for maps