Make .pb.go deterministic by deterministic marshal
We have large diffs in generated .pb.go files even when there is no change in the protobuf itself. Example:

https://github.com/openbank/openbank/commit/1535696c0f58698898b266dee9bbd4ff144dc874#diff-790a2e5fdb0867ef31e92c5b78946f3c

I found out it's actually caused by non-deterministic marshaling of maps; by default, proto.Marshal is non-deterministic with maps, and grpc uses maps to save response types.

If you actually un-gzip the byte array and diff the filedescriptor, the only changes are in the order of the maps in GRPC.

You can marshal deterministcally using proto.Buffer, so I added a helper func that does that