loader: Use Go Overlays to parse Gunk-only packages
The previous strategy involved generating fake Go files and writing them
to disks before parsing the packages to make the Go toolchain happy.
This commit uses Go overlays to fake Gunk packages as Go packages
instead.

An additional benefit that comes with this is that fake Go files no
longer need to be present in modules that provide Gunk-only packages as
the overlay does not have read-only restrictions in the Go module cache
directory.