loader: remove redundant -: prefixes
Previously loader is wrapping all errors as a packages.Error regardless
of error type as there is no way to check if the error should be
wrapped.

This commit changes error addition to allow for passing the error
directly which enables `addError` to check for specific error types and
retrieve information as applicable.