feat(generate): add doc generation mode
this mode will allow generation of the gunk files in JSON format with
primarily used when using the results to create autogenerated
documentation. See generate/doc/json.go for the format for the output
format. The packages are grouped into tags, otherwise packages are
outputted as part of the default tag.