generate: Add generate_single option
The generate_single option allows users to specify that a specific
generator should be ran with all packages at once.

This allows generators that require full context and generators that
only output a single file to run correctly.

When the `generate_single` option is used, the output path must be
specified.

Co-authored-by: Hamza Ali <me@hamzantal.pw>